python_opencv/pyproject.toml
[tool.poetry]
name = "music-traffic"
version = "0.1.0"
description = ""
authors = ["Edward C. Deaver, IV <EdwardDeaver@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
imageai = "^3.0.3"
pillow = "^10.3.0"
numpy = "^1.26.4"
torch = "^2.2.2"
torchvision = "^0.17.2"
pandas = "^2.2.2"
ultralytics = "^8.1.47"
opencv-python-headless = "^4.9.0.80"
pyo = "^1.0.5"
python-rtmidi = "^1.5.8"
mido = "^1.3.2"
pika = "^1.3.2"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
